Added app toolbar and settings dialog
You can now open app's settings with the gear icon in the toolbar.
The ble scan now uses Settings.retryDelay*0.5 instead of 0 as delay between broadcast events.

+++ app/src/main/java/me/itzvirtual/btguard/db/AppSettings.kt
@@ -9,5 +9,5 @@ data class AppSettings(
 	var devices: MutableList<BluetoothDeviceState> = mutableListOf()
 ) {
 	@Serializable
-	data class Settings(var retryDelay: Int = 2000)
+	data class Settings(var retryDelay: Long = 2000)
 }
